dont remove .gitignore from tmp in tests

+++ test/src/server/routes/images.js
@@ -35,6 +35,7 @@ describe('/api/images', function () {
 				if (files.length > 0) {
 					for (var i = 0; i < files.length; i++) {
 						var filePath = dirPath + '/' + files[i];
+						if(filePath == 'tmp/.gitignore') continue;
 						if (fs.statSync(filePath).isFile())
 							fs.unlinkSync(filePath);
 						else
